backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
blinker==1.7.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==42.0.8
customtkinter==5.2.2
darkdetect==0.8.0
Deprecated==1.2.14
distlib==0.3.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.5
fastapi-sso==0.10.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsspec==2024.10.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.52.0
macholib==1.16.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mdurl==0.1.2
multidict==6.1.0
oauthlib==3.2.2
openai==1.54.1
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-semantic-conventions==0.49b0
orjson==3.10.11
packaging==23.2
pillow==11.3.0
platformdirs==4.0.0
propcache==0.2.0
psycopg2-binary==2.9.11
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PyJWT==2.9.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pypdf==6.5.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
regex==2024.9.11
reportlab==4.4.4
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
rq==2.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.37.2
tiktoken==0.8.0
tokenizers==0.20.3
tqdm==4.66.6
typer==0.12.5
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.22.0
uvloop==0.21.0
virtualenv==20.24.7
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.17.1
zipp==3.20.2
